pyproject.toml
[project]
name = "da-staff-software-engineer-assessment-nn"
version = "0.1.0"
description = "Production-Lite streaming transcript and LLM summarization system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.10.0",
    "structlog>=24.1.0",
    "websockets>=12.0",
    "pyjwt>=2.11.0",
    "redis[hiredis]>=5.0.0",
    "httpx[http2]>=0.26.0",
]

[project.optional-dependencies]
nlp = [
    "spacy>=3.7.0,<3.8.0",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]
